Move paged stash Triton kernels
