+++ megatron/core/transformer/moe/kernels/__init__.py
@@ -1,0 +1,1 @@
+# Copyright (c) 2025 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
